Extract serialport.js out of actions.js
So actions.js contains just the logic regarding Redux actions and
serialport.js just the logic regarding communication with the
serialport.
