+++ src/SidePanel/ToggleLed.jsx
@@ -38,7 +38,7 @@ import React from 'react';
 import { useSelector } from 'react-redux';
 import Button from 'react-bootstrap/Button';
 
-import { toggleLED } from '../actions';
+import { toggleLED } from '../serialport';
 import { getIsConnected } from '../reducer';
 
 export default () => {
